+++ ZGO2/include/model/markov/gomarkovstatus.h
@@ -4,7 +4,6 @@
  * The status of GO Markov repairable system.
  * @author ZHG <CyberZHG@gmail.com>
  */
-#include "bigdecimal.h"
 #include "domitem.h"
 
 class QString;
@@ -14,28 +13,28 @@ class GOMarkovStatus : public DomItem
 public:
     GOMarkovStatus();
     virtual ~GOMarkovStatus();
-    BigDecimal probabilityNormal() const;
-    BigDecimal probabilityBreakdown() const;
-    BigDecimal frequencyBreakdown() const;
-    BigDecimal frequencyRepair() const;
+    double probabilityNormal() const;
+    double probabilityBreakdown() const;
+    double frequencyBreakdown() const;
+    double frequencyRepair() const;
     void setProbabilityNormal(QString value);
     void setProbabilityBreakdown(QString value);
     void setFrequencyBreakdown(QString value);
     void setFrequencyRepair(QString value);
     void setRepairTime(QString value);
-    void setProbabilityNormal(BigDecimal value);
-    void setProbabilityBreakdown(BigDecimal value);
-    void setFrequencyBreakdown(BigDecimal value);
-    void setFrequencyRepair(BigDecimal value);
-    void setRepairTime(BigDecimal value);
+    void setProbabilityNormal(double value);
+    void setProbabilityBreakdown(double value);
+    void setFrequencyBreakdown(double value);
+    void setFrequencyRepair(double value);
+    void setRepairTime(double value);
     void save(QDomDocument &document, QDomElement &root);
     bool tryOpen(QDomElement &root);
 
 protected:
-    BigDecimal _probabilityNormal;
-    BigDecimal _probabilityBreakdown;
-    BigDecimal _frequencyBreakdown;
-    BigDecimal _frequencyRepair;
+    double _probabilityNormal;
+    double _probabilityBreakdown;
+    double _frequencyBreakdown;
+    double _frequencyRepair;
 };
 
 #endif // GOMARKOVSTATUS_H
